refactor(scripts): move pure version logic to src/lib for cross-platform tests
Vitest fails to transform ESM imported from outside the project root
(scripts/) on Windows CI with 'SyntaxError: Invalid or unexpected token'.
The failure predates this enhancement (the original PR head had the same
red windows run) and is specific to out-of-root imports in the test
graph.

Move all pure version/CHANGELOG/tag logic into src/lib/version-bump.mjs
- inside the vitest transform root, with no Node built-ins and no
import.meta.url in the shared module - and turn both CLI scripts into
thin shells importing from it. Unit tests now import only from src/, so
the module transforms identically on macOS, ubuntu, and windows.

Also drop the leftover non-ASCII glyphs from the previous attempt; the
whole versioning dependency graph is now pure ASCII.
