+++ scripts/verify-release-tag.mjs
@@ -14,54 +14,20 @@
  * The tag may be passed as argv[2], or taken from the GITHUB_REF_NAME
  * environment variable when invoked from a GitHub Actions tag push.
  * Exit code 0 = tag matches every version file; 1 = mismatch or bad tag.
+ *
+ * The pure tag/version logic (parseReleaseTag, verifyTagAgainstFiles,
+ * collectFileVersions) lives in src/lib/version-bump.mjs, shared with
+ * scripts/bump-version.mjs and the unit tests.
  */
 
 import fs from 'fs';
 import path from 'path';
 import { fileURLToPath } from 'url';
 import {
-  parseVersion,
-  collectFileVersions,
-  findVersionDrift
-} from './bump-version.mjs';
-
-/**
- * Pure tag validation. Returns { ok, version?, error? }. Accepts any semver
- * `vX.Y.Z` or `vX.Y.Z-<prerelease>` tag (stable and tagged prereleases);
- * rejects build metadata and non-semver shapes.
- */
-export function parseReleaseTag(tag) {
-  const trimmed = String(tag || '').trim();
-  const SEMVER_IDENTIFIER = '(?:[0-9A-Za-z-]+)(?:\\.[0-9A-Za-z-]+)*';
-  const tagPattern = new RegExp(`^v\\d+\\.\\d+\\.\\d+(-${SEMVER_IDENTIFIER})?$`);
-  if (!tagPattern.test(trimmed)) {
-    return {
-      ok: false,
-      error: `Invalid release tag '${trimmed}': expected vX.Y.Z or vX.Y.Z-<prerelease> (e.g. v2.8.0, v2.8.0-beta.1)`
-    };
-  }
-  return { ok: true, tag: trimmed, version: trimmed.slice(1) };
-}
-
-/**
- * Pure sync check: does the tag version match every version file?
- * Returns { ok, mismatches: [{file, expectedTagVersion, fileVersion}] }.
- */
-export function verifyTagAgainstFiles(tag, fileContents) {
-  const parsed = parseReleaseTag(tag);
-  if (!parsed.ok) {
-    return { ok: false, mismatches: [], error: parsed.error };
-  }
-  const versions = collectFileVersions(fileContents);
-  const mismatches = Object.entries(versions)
-    .filter(([, version]) => version !== parsed.version)
-    .map(([file, version]) => ({
-      file,
-      expectedTagVersion: parsed.version,
-      fileVersion: version
-    }));
-  return { ok: mismatches.length === 0, mismatches };
-}
+  parseReleaseTag,
+  verifyTagAgainstFiles,
+  collectFileVersions
+} from '../src/lib/version-bump.mjs';
 
 function main() {
   const args = process.argv.slice(2);
